feat(pricing): select sync source in price settings
