+++ web/src/components/usage/PriceSettingsCard.tsx
@@ -5,18 +5,13 @@ import { Button } from '@/components/ui/Button';
 import { Input } from '@/components/ui/Input';
 import { Modal } from '@/components/ui/Modal';
 import { Select, type SelectOption } from '@/components/ui/Select';
-import { IconCheck, IconCircleAlert, IconRefreshCw } from '@/components/ui/icons';
+import { IconCheck } from '@/components/ui/icons';
 import { useScrollBoundaryContainment } from '@/hooks/useScrollBoundaryContainment';
-import { ApiError } from '@/lib/api';
-import type { ModelPrice, PricingRule, PricingSaveResult, PricingStyle, PricingSyncMatch, PricingSyncPreviewResponse, ReplacePricingRuleInput } from '@/lib/types';
+import type { ModelPrice, PricingRule, PricingSaveResult, PricingStyle, PricingSyncSource, PricingSyncPreviewResponse, ReplacePricingRuleInput } from '@/lib/types';
 import { PriceRulesModal } from './pricing/PriceRulesModal';
+import { PriceSyncPanel } from './pricing/PriceSyncPanel';
+import { formatDisplayName, priceToInputValue, pricingDraftToModelPrice, pricingStyleOptions } from './pricing/pricingDrafts';
 import styles from '@/pages/UsagePage.module.scss';
-
-const formatDisplayName = (value: string): string => {
-  const normalized = value.trim();
-  if (!normalized) return '-';
-  return normalized;
-};
 
 const modelNameCollator = new Intl.Collator('en', {
   numeric: true,
@@ -45,212 +40,12 @@ export interface PriceSettingsCardProps {
   onRulesLoad?: (model: string) => Promise<PricingRule[] | null>;
   onRulesSave?: (model: string, rules: ReplacePricingRuleInput[]) => Promise<PricingRule[] | null>;
   onSyncPricesChange?: (prices: Record<string, ModelPrice>) => Promise<PricingSaveResult>;
-  onSyncPreview?: () => Promise<PricingSyncPreviewResponse>;
+  onSyncPreview?: (source: PricingSyncSource, signal?: AbortSignal) => Promise<PricingSyncPreviewResponse>;
   onNotice?: (kind: 'success' | 'info' | 'error', message: string) => void;
   loading?: boolean;
 }
 
 const emptyPricingRules = async (): Promise<PricingRule[]> => [];
-
-export interface PricingSyncDraft {
-  model: string;
-  matchedModel: string;
-  matchType: string;
-  sourceProviderId: string;
-  sourceProviderName: string;
-  selected: boolean;
-  style: PricingStyle;
-  prompt: string;
-  completion: string;
-  cacheRead: string;
-  cacheWrite: string;
-  multiplier: string;
-  saveStatus?: 'failed';
-  saveError?: string;
-}
-
-export interface PricingDraftInput {
-  style: PricingStyle;
-  prompt: string;
-  completion: string;
-  cacheRead: string;
-  cacheWrite: string;
-  multiplier: string;
-}
-
-const parsePriceValue = (value: string): number | null => {
-  const parsed = Number(value);
-  return Number.isFinite(parsed) && parsed >= 0 ? parsed : null;
-};
-
-const parseMultiplierValue = (value: string): number | null => {
-  if (value.trim() === '') return 1;
-  const parsed = Number(value);
-  return Number.isFinite(parsed) && parsed >= 0 ? parsed : null;
-};
-
-const parseOptionalCachePriceValue = (value: string): number | null => (
-  value.trim() === '' ? 0 : parsePriceValue(value)
-);
-
-const priceToInputValue = (value: number | undefined): string => (
-  typeof value === 'number' && Number.isFinite(value) ? value.toString() : ''
-);
-
-const normalizePricingStyle = (style: PricingStyle | string | undefined): PricingStyle => (
-  style === 'claude' ? 'claude' : 'openai'
-);
-
-export const syncMatchToDraft = (match: PricingSyncMatch, existingPrice?: ModelPrice): PricingSyncDraft => ({
-  model: match.model,
-  matchedModel: match.matched_model,
-  matchType: match.match_type,
-  sourceProviderId: match.source_provider_id,
-  sourceProviderName: match.source_provider_name,
-  selected: true,
-  style: normalizePricingStyle(match.pricing_style),
-  prompt: priceToInputValue(match.prompt_price_per_1m),
-  completion: priceToInputValue(match.completion_price_per_1m),
-  cacheRead: priceToInputValue(match.cache_read_price_per_1m),
-  cacheWrite: priceToInputValue(match.cache_write_price_per_1m),
-  multiplier: priceToInputValue(existingPrice?.multiplier ?? 1),
-});
-
-export const pricingDraftToModelPrice = (draft: PricingDraftInput): ModelPrice | null => {
-  const prompt = parsePriceValue(draft.prompt);
-  const completion = parsePriceValue(draft.completion);
-  if (prompt === null || completion === null) return null;
-  const cacheRead = parseOptionalCachePriceValue(draft.cacheRead);
-  const cacheWrite = parseOptionalCachePriceValue(draft.cacheWrite);
-  const multiplier = parseMultiplierValue(draft.multiplier);
-  if (cacheRead === null || cacheWrite === null || multiplier === null) return null;
-  return {
-    style: draft.style,
-    prompt,
-    completion,
-    cacheRead,
-    cacheWrite,
-    multiplier,
-  };
-};
-
-export const syncDraftToModelPrice = (draft: PricingSyncDraft): ModelPrice | null => (
-  pricingDraftToModelPrice(draft)
-);
-
-export const markPricingSyncFailures = (
-  drafts: PricingSyncDraft[],
-  result: PricingSaveResult,
-): PricingSyncDraft[] => {
-  const failedByModel = new Map(result.failures.map((failure) => [failure.model, failure.message]));
-  const successModels = new Set(result.successModels);
-  return drafts.map((draft) => {
-    const failureMessage = failedByModel.get(draft.model);
-    if (failureMessage !== undefined) {
-      return {
-        ...draft,
-        selected: true,
-        saveStatus: 'failed',
-        saveError: failureMessage,
-      };
-    }
-    if (successModels.has(draft.model)) {
-      return {
-        ...draft,
-        selected: false,
-        saveStatus: undefined,
-        saveError: undefined,
-      };
-    }
-    return {
-      ...draft,
-      saveStatus: undefined,
-      saveError: undefined,
-    };
-  });
-};
-
-export const notifyPricingSyncUnexpectedError = (
-  error: unknown,
-  t: (key: string) => string,
-  onNotice: PriceSettingsCardProps['onNotice'],
-) => {
-  if (error instanceof ApiError && error.status === 504) {
-    onNotice?.('error', t('usage_stats.model_price_sync_timeout'));
-    return;
-  }
-
-  const message = error instanceof Error ? error.message : '';
-  onNotice?.(
-    'error',
-    `${t('usage_stats.model_price_sync_failed')}${message ? `: ${message}` : ''}`,
-  );
-};
-
-export const notifyPricingSyncFailures = (
-  result: PricingSaveResult,
-  t: (key: string, options?: { success: number; failed: number }) => string,
-  onNotice: PriceSettingsCardProps['onNotice'],
-) => {
-  if (result.failures.length === 0) return;
-  const summary = t('usage_stats.model_price_sync_apply_partial', {
-    success: result.successModels.length,
-    failed: result.failures.length,
-  });
-  const detail = result.failures.find((failure) => failure.message.trim())?.message.trim() ?? '';
-  onNotice?.(
-    result.successModels.length > 0 ? 'info' : 'error',
-    `${summary}${detail ? `: ${detail}` : ''}`,
-  );
-};
-
-export interface SelectedSyncPrices {
-  selectedDrafts: PricingSyncDraft[];
-  prices: Record<string, ModelPrice>;
-  invalidModel: string | null;
-}
-
-export const buildSelectedSyncPrices = (drafts: PricingSyncDraft[]): SelectedSyncPrices => {
-  const selectedDrafts = drafts.filter((draft) => draft.selected);
-  const prices: Record<string, ModelPrice> = {};
-  for (const draft of selectedDrafts) {
-    const price = syncDraftToModelPrice(draft);
-    if (!price) {
-      return { selectedDrafts, prices: {}, invalidModel: draft.model };
-    }
-    prices[draft.model] = price;
-  }
-  return { selectedDrafts, prices, invalidModel: null };
-};
-
-export const saveSyncDraftsWithSingleModelCallback = async (
-  selectedDrafts: PricingSyncDraft[],
-  prices: Record<string, ModelPrice>,
-  onPriceSave: PriceSettingsCardProps['onPriceSave'],
-): Promise<PricingSaveResult> => {
-  const settled = await Promise.all(selectedDrafts.map(async (draft) => {
-    try {
-      await Promise.resolve(onPriceSave(draft.model, prices[draft.model]));
-      return { model: draft.model, ok: true as const };
-    } catch (error) {
-      return {
-        model: draft.model,
-        ok: false as const,
-        message: error instanceof Error ? error.message : String(error),
-        error,
-      };
-    }
-  }));
-
-  return settled.reduce<PricingSaveResult>((result, item) => {
-    if (item.ok) {
-      result.successModels.push(item.model);
-    } else {
-      result.failures.push({ model: item.model, message: item.message, error: item.error });
-    }
-    return result;
-  }, { successModels: [], failures: [] });
-};
 
 const notifyPricingPersistenceError = (
   error: unknown,
@@ -260,11 +55,6 @@ const notifyPricingPersistenceError = (
   const message = error instanceof Error ? error.message : '';
   onNotice?.('error', `${fallbackMessage}${message ? `: ${message}` : ''}`);
 };
-
-const pricingStyleOptions = (t: (key: string) => string): SelectOption[] => [
-  { value: 'openai', label: t('usage_stats.model_price_style_openai') },
-  { value: 'claude', label: t('usage_stats.model_price_style_claude') },
-];
 
 export const buildPricingModelOptions = (
   modelNames: string[],
@@ -331,12 +121,6 @@ export function PriceSettingsCard({
   const [deleteModel, setDeleteModel] = useState<string | null>(null);
   const [deleteSaving, setDeleteSaving] = useState(false);
   const [rulesModel, setRulesModel] = useState<string | null>(null);
-
-  const [syncOpen, setSyncOpen] = useState(false);
-  const [syncLoading, setSyncLoading] = useState(false);
-  const [syncApplying, setSyncApplying] = useState(false);
-  const [syncPreview, setSyncPreview] = useState<PricingSyncPreviewResponse | null>(null);
-  const [syncDrafts, setSyncDrafts] = useState<PricingSyncDraft[]>([]);
 
   const closeEditModal = () => {
     if (!editSaving) {
@@ -454,76 +238,6 @@ export function PriceSettingsCard({
     }
   };
 
-  const handleOpenSyncPreview = async () => {
-    if (!onSyncPreview || syncLoading) return;
-    setSyncLoading(true);
-    try {
-      const preview = await onSyncPreview();
-      const drafts = (preview.matches ?? []).map((match) => syncMatchToDraft(match, modelPrices[match.model]));
-      setSyncPreview({
-        ...preview,
-        matches: preview.matches ?? [],
-        unmatched_models: preview.unmatched_models ?? [],
-      });
-      setSyncDrafts(drafts);
-      setSyncOpen(true);
-      if (drafts.length === 0) {
-        onNotice?.('info', t('usage_stats.model_price_sync_no_matches'));
-      }
-    } catch (error) {
-      notifyPricingSyncUnexpectedError(error, t, onNotice);
-    } finally {
-      setSyncLoading(false);
-    }
-  };
-
-  const handleUpdateSyncDraft = (index: number, patch: Partial<PricingSyncDraft>) => {
-    const clearsFailure = Object.keys(patch).some((key) => key !== 'selected');
-    setSyncDrafts((current) => current.map((draft, draftIndex) => (
-      draftIndex === index
-        ? {
-          ...draft,
-          ...patch,
-          ...(clearsFailure ? { saveStatus: undefined, saveError: undefined } : {}),
-        }
-        : draft
-    )));
-  };
-
-  const handleSetAllSyncDrafts = (selected: boolean) => {
-    setSyncDrafts((current) => current.map((draft) => ({ ...draft, selected })));
-  };
-
-  const handleApplySyncDrafts = async () => {
-    const { selectedDrafts, prices: syncPrices, invalidModel } = buildSelectedSyncPrices(syncDrafts);
-    if (selectedDrafts.length === 0) {
-      onNotice?.('error', t('usage_stats.model_price_sync_none_selected'));
-      return;
-    }
-    if (invalidModel !== null) {
-      onNotice?.('error', t('usage_stats.model_price_sync_invalid', { model: formatDisplayName(invalidModel) }));
-      return;
-    }
-
-    setSyncApplying(true);
-    try {
-      const result = onSyncPricesChange
-        ? await onSyncPricesChange(syncPrices)
-        : await saveSyncDraftsWithSingleModelCallback(selectedDrafts, syncPrices, onPriceSave);
-      setSyncDrafts((current) => markPricingSyncFailures(current, result));
-      if (result.failures.length === 0) {
-        onNotice?.('success', t('usage_stats.model_price_sync_apply_success', { count: result.successModels.length }));
-        setSyncOpen(false);
-        return;
-      }
-      notifyPricingSyncFailures(result, t, onNotice);
-    } catch (error) {
-      notifyPricingSyncUnexpectedError(error, t, onNotice);
-    } finally {
-      setSyncApplying(false);
-    }
-  };
-
   const options = useMemo(
     () => buildPricingModelOptions(
       modelNames,
@@ -540,11 +254,6 @@ export function PriceSettingsCard({
     [modelPrices]
   );
   useScrollBoundaryContainment(pricesGridRef, sortedModelPrices.length > 0);
-  const selectedSyncCount = useMemo(
-    () => syncDrafts.filter((draft) => draft.selected).length,
-    [syncDrafts]
-  );
-
   return (
     <>
       <Card
@@ -557,21 +266,9 @@ export function PriceSettingsCard({
             <div className={styles.hint}>{t('common.loading')}</div>
           ) : (
             <>
-              {onSyncPreview && (
-                <div className={styles.pricingToolbar}>
-                  <div className={styles.pricingToolbarMeta}>
-                    <span>{t('usage_stats.model_price_sync_source')}: Models.dev</span>
-                  </div>
-                  <Button
-                    variant="secondary"
-                    appearance="action"
-                    onClick={() => void handleOpenSyncPreview()}
-                    loading={syncLoading}
-                  >
-                    <IconRefreshCw size={14} />
-                    {t('usage_stats.model_price_sync')}
-                  </Button>
-                </div>
+              {onSyncPreview && onSyncPricesChange && (
+                <PriceSyncPanel modelPrices={modelPrices} onSyncPreview={onSyncPreview}
+                  onSyncPricesChange={onSyncPricesChange} onNotice={onNotice} />
               )}
               <div className={styles.priceForm}>
                 <div className={styles.formRow}>
@@ -839,219 +536,6 @@ export function PriceSettingsCard({
         </p>
       </Modal>
 
-      <Modal
-        open={syncOpen}
-        title={t('usage_stats.model_price_sync_title')}
-        onClose={() => {
-          if (!syncApplying) {
-            setSyncOpen(false);
-          }
-        }}
-        closeDisabled={syncApplying}
-        footer={
-          <div className={styles.priceActions}>
-            <Button
-              variant="secondary"
-              appearance="action"
-              onClick={() => setSyncOpen(false)}
-              disabled={syncApplying}
-            >
-              {t('common.cancel')}
-            </Button>
-            <Button
-              variant="primary"
-              appearance="action"
-              onClick={() => void handleApplySyncDrafts()}
-              loading={syncApplying}
-              disabled={selectedSyncCount === 0}
-            >
-              {t('usage_stats.model_price_sync_update_selected', { count: selectedSyncCount })}
-            </Button>
-          </div>
-        }
-        width={940}
-      >
-        <div className={styles.syncModalBody}>
-          <div className={styles.syncSummaryRow}>
-            <span>
-              {t('usage_stats.model_price_sync_source')}: {syncPreview?.source || 'Models.dev'}
-            </span>
-            <span>
-              {t('usage_stats.model_price_sync_matched')}: {syncDrafts.length}
-            </span>
-            <span>
-              {t('usage_stats.model_price_sync_unmatched')}: {syncPreview?.unmatched_models?.length ?? 0}
-            </span>
-          </div>
-
-          {syncDrafts.length > 0 ? (
-            <>
-              <div className={styles.syncBatchActions}>
-                <Button
-                  variant="secondary"
-                  size="sm"
-                  appearance="action"
-                  onClick={() => handleSetAllSyncDrafts(true)}
-                  disabled={syncApplying}
-                >
-                  {t('usage_stats.model_price_sync_select_all')}
-                </Button>
-                <Button
-                  variant="secondary"
-                  size="sm"
-                  appearance="action"
-                  onClick={() => handleSetAllSyncDrafts(false)}
-                  disabled={syncApplying}
-                >
-                  {t('usage_stats.model_price_sync_select_none')}
-                </Button>
-              </div>
-
-              <div className={styles.syncDraftList}>
-                {syncDrafts.map((draft, index) => {
-                  const existing = Boolean(modelPrices[draft.model]);
-                  const failed = draft.saveStatus === 'failed';
-                  const failureLabel = t('usage_stats.model_price_sync_failed_label', { model: formatDisplayName(draft.model) });
-                  return (
-                    <div
-                      key={`${draft.model}-${draft.matchedModel}`}
-                      className={`${styles.syncDraftItem} ${failed ? styles.syncDraftItemFailed : ''}`}
-                    >
-                      <label className={styles.syncDraftCheck}>
-                        <input
-                          type="checkbox"
-                          checked={draft.selected}
-                          disabled={syncApplying}
-                          onChange={(event) => handleUpdateSyncDraft(index, { selected: event.target.checked })}
-                          aria-label={t('usage_stats.model_price_sync_toggle', { model: formatDisplayName(draft.model) })}
-                        />
-                      </label>
-                      <div className={styles.syncDraftContent}>
-                        <div className={styles.syncDraftHeader}>
-                          <div className={styles.syncDraftModelBlock}>
-                            <span className={styles.priceModel}>{formatDisplayName(draft.model)}</span>
-                            <span className={styles.syncDraftMatched}>
-                              {t('usage_stats.model_price_sync_matched_model', { model: formatDisplayName(draft.matchedModel) })}
-                            </span>
-                            <span className={styles.syncDraftMatched}>
-                              {t('usage_stats.model_price_sync_provider', {
-                                provider: formatDisplayName(draft.sourceProviderName || draft.sourceProviderId),
-                                id: formatDisplayName(draft.sourceProviderId),
-                              })}
-                            </span>
-                          </div>
-                          <div className={styles.syncDraftBadges}>
-                            {failed && (
-                              <span
-                                className={styles.syncDraftFailureIcon}
-                                role="img"
-                                aria-label={failureLabel}
-                                title={draft.saveError || failureLabel}
-                              >
-                                <IconCircleAlert size={13} />
-                              </span>
-                            )}
-                            <span>{draft.matchType}</span>
-                            {existing && <span>{t('usage_stats.model_price_sync_existing')}</span>}
-                          </div>
-                        </div>
-                        <div className={styles.syncDraftGrid}>
-                          <div className={styles.formField}>
-                            <label>{t('usage_stats.model_price_style')}</label>
-                            <Select
-                              value={draft.style}
-                              options={styleOptions}
-                              onChange={(value) => handleUpdateSyncDraft(index, { style: value === 'claude' ? 'claude' : 'openai' })}
-                              disabled={syncApplying}
-                              className={styles.usagePillControl}
-                            />
-                          </div>
-                          <div className={styles.formField}>
-                            <label>{t('usage_stats.model_price_prompt')} ($/1M)</label>
-                            <Input
-                              type="number"
-                              value={draft.prompt}
-                              onChange={(event) => handleUpdateSyncDraft(index, { prompt: event.target.value })}
-                              placeholder="0.00"
-                              step="0.0001"
-                              disabled={syncApplying}
-                              className={styles.usagePillControl}
-                            />
-                          </div>
-                          <div className={styles.formField}>
-                            <label>{t('usage_stats.model_price_completion')} ($/1M)</label>
-                            <Input
-                              type="number"
-                              value={draft.completion}
-                              onChange={(event) => handleUpdateSyncDraft(index, { completion: event.target.value })}
-                              placeholder="0.00"
-                              step="0.0001"
-                              disabled={syncApplying}
-                              className={styles.usagePillControl}
-                            />
-                          </div>
-                          <div className={styles.formField}>
-                            <label>{t('usage_stats.model_price_cache_read')} ($/1M)</label>
-                            <Input
-                              type="number"
-                              value={draft.cacheRead}
-                              onChange={(event) => handleUpdateSyncDraft(index, { cacheRead: event.target.value })}
-                              placeholder="0.00"
-                              step="0.0001"
-                              disabled={syncApplying}
-                              className={styles.usagePillControl}
-                            />
-                          </div>
-                          <div className={styles.formField}>
-                            <label>{t('usage_stats.model_price_cache_write')} ($/1M)</label>
-                            <Input
-                              type="number"
-                              value={draft.cacheWrite}
-                              onChange={(event) => handleUpdateSyncDraft(index, { cacheWrite: event.target.value })}
-                              placeholder="0.00"
-                              step="0.0001"
-                              disabled={syncApplying}
-                              className={styles.usagePillControl}
-                            />
-                          </div>
-                          <div className={styles.formField}>
-                            <label>{t('usage_stats.model_price_multiplier')}</label>
-                            <Input
-                              type="number"
-                              value={draft.multiplier}
-                              onChange={(event) => handleUpdateSyncDraft(index, { multiplier: event.target.value })}
-                              placeholder="1"
-                              step="0.0001"
-                              min="0"
-                              disabled={syncApplying}
-                              className={styles.usagePillControl}
-                            />
-                          </div>
-                        </div>
-                      </div>
-                    </div>
-                  );
-                })}
-              </div>
-            </>
-          ) : (
-            <div className={styles.hint}>{t('usage_stats.model_price_sync_no_matches')}</div>
-          )}
-
-          {(syncPreview?.unmatched_models?.length ?? 0) > 0 && (
-            <details className={styles.syncUnmatched}>
-              <summary>
-                {t('usage_stats.model_price_sync_unmatched')}: {syncPreview?.unmatched_models.length}
-              </summary>
-              <div className={styles.syncUnmatchedList}>
-                {syncPreview?.unmatched_models.map((model) => (
-                  <span key={model}>{formatDisplayName(model)}</span>
-                ))}
-              </div>
-            </details>
-          )}
-        </div>
-      </Modal>
     </>
   );
 }
