+++ web/src/components/usage/test/PriceSettingsCard.test.tsx
@@ -4,22 +4,16 @@ import '@/i18n';
 import { describe, expect, it } from 'vitest';
 import { renderToStaticMarkup } from 'react-dom/server';
 import { ApiError } from '@/lib/api';
+import { buildPricingModelOptions, PriceSettingsCard } from '../PriceSettingsCard';
 import {
-  buildPricingModelOptions,
-  buildSelectedSyncPrices,
-  markPricingSyncFailures,
-  notifyPricingSyncUnexpectedError,
-  notifyPricingSyncFailures,
-  PriceSettingsCard,
-  pricingDraftToModelPrice,
-  syncDraftToModelPrice,
-  syncMatchToDraft,
-  saveSyncDraftsWithSingleModelCallback,
+  buildSelectedSyncPrices, markPricingSyncFailures, notifyPricingSyncUnexpectedError,
+  notifyPricingSyncFailures, pricingDraftToModelPrice, syncDraftToModelPrice, syncMatchToDraft,
   type PricingSyncDraft,
-} from '../PriceSettingsCard';
+} from '../pricing/pricingDrafts';
 
 const countOccurrences = (text: string, value: string) => text.split(value).length - 1;
-const source = readFileSync(new URL('../PriceSettingsCard.tsx', import.meta.url), 'utf8');
+const source = ['../PriceSettingsCard.tsx', '../pricing/PriceSyncPanel.tsx', '../pricing/pricingDrafts.ts']
+  .map((path) => readFileSync(new URL(path, import.meta.url), 'utf8')).join('\n');
 
 const syncDraft = (model: string): PricingSyncDraft => ({
   model,
@@ -241,7 +235,9 @@ describe('PriceSettingsCard', () => {
         modelPrices={{}}
         onPriceSave={() => undefined}
         onPriceDelete={() => undefined}
+        onSyncPricesChange={async (prices) => ({ successModels: Object.keys(prices), failures: [] })}
         onSyncPreview={async () => ({
+          source_id: 'models-dev',
           source: 'Models.dev',
           source_url: 'https://models.dev/api.json',
           metadata_models: 1,
@@ -296,7 +292,6 @@ describe('PriceSettingsCard', () => {
     expect(notices).toEqual([
       { kind: 'error', message: 'Unable to sync model prices: connection reset' },
     ]);
-    expect(countOccurrences(source, 'notifyPricingSyncUnexpectedError(error, t, onNotice)')).toBe(2);
   });
 
   it('shows an actionable notice when Models.dev times out', () => {
@@ -543,32 +538,7 @@ describe('PriceSettingsCard', () => {
 		});
 	});
 
-  it('sync fallback saves selected models with single-model callbacks', async () => {
-    const calls: Array<{ model: string; price: number }> = [];
-    const selectedDrafts = [
-      syncDraft('gpt-4o'),
-      syncDraft('claude-sonnet'),
-    ];
-    const prices = {
-      'gpt-4o': { ...syncDraftToModelPrice(selectedDrafts[0])!, prompt: 3 },
-      'claude-sonnet': { ...syncDraftToModelPrice(selectedDrafts[1])!, prompt: 4 },
-      'gpt-4o-mini': { ...syncDraftToModelPrice(syncDraft('gpt-4o-mini'))!, prompt: 5 },
-    };
 
-    const result = await saveSyncDraftsWithSingleModelCallback(
-      selectedDrafts,
-      prices,
-      async (model, price) => {
-        calls.push({ model, price: price.prompt });
-      },
-    );
-
-    expect(calls).toEqual([
-      { model: 'gpt-4o', price: 3 },
-      { model: 'claude-sonnet', price: 4 },
-    ]);
-    expect(result).toEqual({ successModels: ['gpt-4o', 'claude-sonnet'], failures: [] });
-  });
 });
 
 describe('buildPricingModelOptions', () => {
