+++ web/src/i18n/index.ts
@@ -676,7 +676,7 @@ const resources = {
         model_price_sync_no_matches: 'No matched model prices found.',
         model_price_sync_none_selected: 'Select at least one matched model to update.',
         model_price_sync_failed: 'Unable to sync model prices',
-        model_price_sync_timeout: 'Models.dev connection timed out. Check the Keeper server network connection and try again.',
+        model_price_sync_timeout: '{{source}} connection timed out. Check the Keeper server network connection and try again.',
         model_price_sync_invalid: 'Check the price values for {{model}} before updating.',
         model_price_sync_apply_success: 'Updated {{count}} model prices.',
         model_price_sync_apply_partial: 'Updated {{success}} model prices, {{failed}} failed.',
@@ -1522,7 +1522,7 @@ const resources = {
         model_price_sync_no_matches: '没有找到可匹配的模型价格。',
         model_price_sync_none_selected: '请至少选择一个匹配模型再更新。',
         model_price_sync_failed: '无法同步模型价格',
-        model_price_sync_timeout: '连接 Models.dev 超时，请检查 Keeper 服务端网络后重试。',
+        model_price_sync_timeout: '连接 {{source}} 超时，请检查 Keeper 服务端网络后重试。',
         model_price_sync_invalid: '请先检查 {{model}} 的价格值。',
         model_price_sync_apply_success: '已更新 {{count}} 个模型价格。',
         model_price_sync_apply_partial: '已更新 {{success}} 个模型价格，{{failed}} 个失败。',
@@ -2368,7 +2368,7 @@ const resources = {
         model_price_sync_no_matches: '沒有找到可匹配的模型價格。',
         model_price_sync_none_selected: '請至少選擇一個匹配模型再更新。',
         model_price_sync_failed: '無法同步模型價格',
-        model_price_sync_timeout: '連線 Models.dev 逾時，請檢查 Keeper 服務端網路後再試。',
+        model_price_sync_timeout: '連線 {{source}} 逾時，請檢查 Keeper 服務端網路後再試。',
         model_price_sync_invalid: '請先檢查 {{model}} 的價格值。',
         model_price_sync_apply_success: '已更新 {{count}} 個模型價格。',
         model_price_sync_apply_partial: '已更新 {{success}} 個模型價格，{{failed}} 個失敗。',
